backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
Pillow==11.2.1
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
django-filter==23.1
reportlab==4.4.1